requirements.txt
aiofiles==0.6.0
appdirs==1.4.4
astroid==2.4.2
astunparse==1.6.3
black==20.8b1
bravado==10.6.2
bravado-core==5.17.0
bunch==1.0.1
cachetools==4.1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
ConfigArgParse==1.4
cycler==0.10.0
Cython==0.29.19
decorator==4.4.2
fastapi==0.63.0
feedparser==6.0.2
flake8==3.8.2
Flask==1.1.2
Flask-BasicAuth==0.2.0
gast==0.3.3
gevent==21.1.2
geventhttpclient==1.4.4
gitdb==4.0.5
GitPython==3.1.3
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.17
h11==0.12.0
h5py==2.10.0
idna==2.10
imagecodecs==2020.2.18
imageio==2.8.0
imgaug==0.4.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.3
joblib==1.0.1
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
locust==1.4.4
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
msgpack==1.0.0
msgpack-python==0.5.6
mypy-extensions==0.4.3
neovim==0.3.1
networkx==2.4
nltk==3.6.2
opencv-python==4.2.0.34
packaging==20.4
pandas==1.0.4
parso==0.7.1
pathspec==0.8.0
protobuf==3.12.1
psutil==5.8.0
py3nvml==0.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydantic==1.8.1
pyflakes==2.2.0
pylint==2.6.0
pynvim==0.4.2
pyparsing==2.4.7
pyrsistent==0.16.0
python-multipart==0.0.5
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
regex==2020.10.28
requests==2.25.1
requests-oauthlib==1.3.0
rfc3987==1.3.8
rsa==4.0
scipy==1.4.1
sgmllib3k==1.0.0
Shapely==1.7.0
six==1.15.0
smmap==3.0.4
starlette==0.13.6
strict-rfc3339==0.7
swagger-spec-validator==2.7.0
tensorboard-plugin-wit==1.6.0.post3
termcolor==1.1.0
tifffile==2020.5.11
toml==0.10.2
tqdm==4.46.1
typed-ast==1.4.1
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.4
uvicorn==0.13.4
webcolors==1.11.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
yapf==0.30.0
zope.event==4.5.0
zope.interface==5.4.0
